backend/requirements.txt
flask
requests
lighter-v1-python
apscheduler
flask-sqlalchemy
flask-cors

hyperliquid-python-sdk
